Split PyProxy method definitions into separate pyproxy.js file (using a bit of macro hackery).
